V1.0: First draft of DFS
Completed the writing of DFS and all the necessary comments. All the tests given in the assignment have been passed, while the tests from the "autograder.py" file result in 0/3 points. This is likely due to the fact that when writing non-iterative DFS, I handle the path incorrectly. I will make sure to correct this later.