V2.0: First draft of BFS
Completed the writing of BFS and all the necessary comments. All the tests given in the assignment have been passed, while the tests from the "autograder.py" file result in 0/3 points. Like DFS I insert a node that doesn't need to be visited for the solution. I will change it in the future update.
Note: The BFS code is equal to DFS, but use a queue instead of a stack. So the code and the comments are most likely to be the same.