V4.3: Completed UCS
The writing of UCS is completed, and it successfully passes the autograder with a full score. Additionally, unnecessary lines of code have been removed for make it cleaner to read, and the comments have been adjusted accordingly. This time is not a simple copy and past of algorithm, but it nedded a good implementation for reconize path with less cost.