V4.4: Completed A*
The writing of A* is completed, and it successfully passes the autograder with a full score and comments have been adjusted accordingly. Additionally, the readability of the algorithm due to the complexity of the code has been reduced by removing additional lines of code and grouping instructions. The most significant change happend to be the removal of the outer for loop; now, the entire algorithm takes place within the while loop. The next commit will focus on consolidating the code and comments in the file.